Refactor initial configuration functionality
Moved initial configuration functionality to a new `CreateInitialConfig` namespace. Updated `CreateInitialConfigCommand` and its handler to implement the `ICommand<Guid>` interface and handle creation logic with error handling. Refactored `InitialConfigCreatedDomainEventHandler` to manage event notifications. Removed unused directives and old namespace references for cleaner code.
