+++ src/EnvironmentGateway/EnvironmentGateway.Application/GatewayConfig/CreateInitialConfig/CreateInitialConfigCommand.cs
@@ -1,5 +1,5 @@
 ﻿using EnvironmentGateway.Application.Abstractions.Messaging;
 
-namespace EnvironmentGateway.Application.GatewayConfig.InitialConfig;
+namespace EnvironmentGateway.Application.GatewayConfig.CreateInitialConfig;
 
 public record CreateInitialConfigCommand(string Name) : ICommand<Guid>;
